Add files via upload
Observations and Insights.
a)as from the data set we can clearly see that graduate ,phd , postgraduate folks were in the green cluster zone 1 and were high on the spending score range of 80 and 100 with low remuneration
b) similarly in black cluster zone 3 major number of people were phd followed by graduate and were in the spending score of 20 -40 with remunartion.
c) people with basic remunartion were scattred all acorss the cluster zones similarly like diploma holders.
few diploma holders were in 40 -60 range on remunartion and red zone 0 cluster
d) overall majority of people were graduate and in red zone cluster 0
e)we can also see that maximum number of people were in zone 0 as observed in the predicted class. with 774 in zone 0, 356 in zone 2 followed by 330 in zone 4.
f) people with basic eductation were very less in and the clusters.